Added 0xf000 instruction; minor changes. See desc
This instruction is part of the CHIP-XO extension and enables 'I' to be
loaded with a 16bit address, instead of the usual 12bit (Therefore,
allowing access to XO's extended memory size of 64kb).
Renamed instruction description for 'SUB (x-=y)' and 'SUB (y-=x)'
to 'SUBX' and 'SUBY' respectively.
Added 'MEM_GET_WORD' macro to extract a 16bit value (word) from memory.